+++ src/jira_test_framework/ui/login_page.py
@@ -4,7 +4,7 @@ from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 
-from src import logger
+from src.logger import logger
 
 
 class LoginPage():
