+++ tests/ui/driver_factory.py
@@ -5,4 +5,5 @@ def get_chrome_driver():
     # this "option" because old writing is deprecated & to cancel an error of "USB: usb_device_handle..." 19-2-23
     options = webdriver.ChromeOptions()
     options.add_experimental_option('excludeSwitches', ['enable-logging'])
+    options.add_argument("--start-maximized")
     return webdriver.Chrome(executable_path='C:\Drivers\chromedriver\chromedriver.exe', options=options)
