+++ tests/ui/login_test.py
@@ -5,14 +5,15 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 
 from src.jira_test_framework.ui.login_page import LoginPage
-from tests import logger
-from tests.ui.driver_factory import get_chrome_driver
+from src.logger import logger
+from tests.ui.constants import INVALID_USERNAME, INVALID_PASSWORD, USERNAME
 
-# I'll take it out to constant file if it will be necessary
-USERNAME = "itzikv3@gmail.com"
-PASSWORD = "itzikpass1212"
-INVALID_USERNAME = "inavlid_username"
-INVALID_PASSWORD = "121212"
+from tests.ui.driver_factory import get_chrome_driver
+from tests.ui.ui_utils import login
+
+your_work_class_name_locator = "css-kwc091"
+continue_xpath_locator = "//span[text()='Continue']"
+help_logging_in_xpath_locator = "//a[text()='logging in']"
 
 
 # @pytest.fixture it's like before test - return driver
@@ -24,12 +25,9 @@ def driver():
 
 
 def test_valid_ui_login(driver):
-    login_page = LoginPage(driver)
-    login_page.login(USERNAME, PASSWORD)
-
-    your_work_css_locator = "css-kwc091"
+    login(driver)
     your_work_element = WebDriverWait(driver, 20).until(
-        EC.visibility_of_element_located((By.CLASS_NAME, your_work_css_locator)))
+        EC.visibility_of_element_located((By.CLASS_NAME, your_work_class_name_locator)))
     assert your_work_element.text == "Your work"
     logger.info("login with ui successfully")
 
@@ -38,7 +36,6 @@ def test_invalid_username_ui_login(driver):
     login_page = LoginPage(driver)
     login_page.set_username(INVALID_USERNAME)
 
-    continue_xpath_locator = "//span[text()='Continue']"
     continue_element = WebDriverWait(driver, 20).until(
         EC.visibility_of_element_located((By.XPATH, continue_xpath_locator)))
 
@@ -50,7 +47,6 @@ def test_invalid_password_login_ui(driver):
     login_page = LoginPage(driver)
     login_page.login(USERNAME, INVALID_PASSWORD)
 
-    help_logging_in_xpath_locator = "//a[text()='logging in']"
     help_logging_in_element = WebDriverWait(driver, 20).until(
         EC.visibility_of_element_located((By.XPATH, help_logging_in_xpath_locator)))
